feat: Generate editable diagram and calculate floats
This Jupyter notebook aims to support you in designing iDesign-informed projects.

You have to enter the project tasks, then the notebook draws an arrow diagram, where arrows represent tasks and nodes represent "task started/ended" events. This might seem counter intuitive, but arrow diagrams shine in case of fan-in followed by fan-out sub-graphs, which are not nearly as readable in node diagrams.

The diagram is generated by a heuristic, which aims to keep the critical path close to the vertical center and then adds additional paths in a non-overlapping way. Because the heuristic does not perfectly avaoid arrow crossings, the diagram is editable and nodes can be moved around.

Vibe-coded.